feat: Integrate ImageNotification extension and Firebase configuration
- Added ImageNotification app extension to the macOS project.
- Configured Firebase in AppDelegate for initialization.
- Updated entitlements to include permissions for Bluetooth, camera, and network access.
- Bumped version in pubspec.yaml to 1.9.17+301.
- Modified Firebase messaging service worker to use alert instead of console log.
- Updated flutter.js version in index.html to 295.

+++ lib/core/base/view/seri_islemleri/seri_listesi/model/seri_detayi_model.dart
@@ -3,8 +3,8 @@ import "../../../../../../view/main_page/alt_sayfalar/stok/base_stok_edit/model/
 final class SeriDetayiModel {
   SeriDetayiModel({this.kalanMiktar, this.hareketMiktari, this.seriList, this.miktarKadarSor});
 
-  int? kalanMiktar;
+  double? kalanMiktar;
   SeriList? seriList;
-  int? hareketMiktari;
+  double? hareketMiktari;
   bool? miktarKadarSor;
 }
